separated every function under helper.ts
