--- src/lib/helper.ts
@@ -1,364 +1,0 @@
-import * as helper from './helper'
-import type {
-  FillLayerSpecification,
-  HeatmapLayerSpecification,
-  LineLayerSpecification,
-  RasterLayerSpecification,
-  SymbolLayerSpecification,
-} from 'maplibre-gl'
-
-import { get } from 'svelte/store'
-import Clipper from 'image-clipper'
-import mime from 'mime'
-import type {
-  BannerMessage,
-  IntervalLegendColorMapRow,
-  Layer,
-  RasterTileMetadata,
-  SpriteIcon,
-  SpriteImage,
-} from './types'
-import { bannerMessages, map } from '$stores'
-import {
-  ClassificationMethodTypes,
-  DEFAULT_TIMEOUT_MS,
-  ErrorMessages,
-  NO_RANDOM_SAMPLING_POINTS,
-  StatusTypes,
-} from './constants'
-import { Jenks } from './jenks'
-import chroma from 'chroma-js'
-
-export const updateParamsInURL = (
-  definition:
-    | RasterLayerSpecification
-    | LineLayerSpecification
-    | FillLayerSpecification
-    | SymbolLayerSpecification
-    | HeatmapLayerSpecification,
-  layerURL: URL,
-  params: Record<string, string>,
-) => {
-  Object.keys(params).forEach((key) => {
-    layerURL.searchParams.set(key, params[key])
-  })
-  const mapStore = get(map)
-  if ('getStyle' in mapStore) {
-    const style = mapStore.getStyle()
-
-    if (style?.sources) {
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      style.sources[definition.source].tiles = [decodeURI(layerURL.toString())]
-      mapStore.setStyle(style)
-    }
-  }
-}
-
-/**
- * Returns a formatted stringified version of json object
- * @param style JSON object
- * @returns string Formatted JSON string
- */
-export const stringifyStyleJSON = (style: JSON): string => {
-  return JSON.stringify(style, null, 4)
-}
-
-export const loadImageToDataUrl = async (url: string): Promise<string> => {
-  const blob = await fetch(url).then((r) => r.blob())
-  return await new Promise((resolve) => {
-    const reader = new FileReader()
-    reader.onload = () => {
-      if (typeof reader.result === 'string') {
-        resolve(reader.result)
-      }
-    }
-    reader.readAsDataURL(blob)
-  })
-}
-
-export const clipSprite = (url: string, id: string, icon: SpriteIcon): Promise<SpriteImage> => {
-  return new Promise((resolve) => {
-    Clipper(url, function () {
-      this.crop(icon.x, icon.y, icon.width, icon.height).toDataURL(function (dataUrl: string) {
-        resolve({
-          src: dataUrl,
-          alt: id,
-        })
-      })
-    })
-  })
-}
-
-/**
- * Returns a json response object from the a fetch of a url
- * @param url The URL to fetch
- * @returns JSON | null
- */
-export async function fetchUrl(url: string) {
-  try {
-    const response = await helper.fetchWithTimeout(url, { timeout: DEFAULT_TIMEOUT_MS })
-    return await response.json()
-  } catch (error) {
-    const bannerErrorMessage: BannerMessage = {
-      type: StatusTypes.DANGER,
-      title: 'Whoops! Something went wrong.',
-      message: ErrorMessages.FETCH_TIMEOUT,
-      error: error,
-    }
-    bannerMessages.update((data) => [...data, bannerErrorMessage])
-    return null
-  }
-}
-
-export async function fetchWithTimeout(resource: string, options = { timeout: DEFAULT_TIMEOUT_MS }) {
-  const { timeout = DEFAULT_TIMEOUT_MS } = options
-  const controller = new AbortController()
-  const id = setTimeout(() => controller.abort(), timeout)
-  const response = await fetch(resource, {
-    ...options,
-    signal: controller.signal,
-  })
-  clearTimeout(id)
-  return response
-}
-
-/**
- * Download a file
- * @param filename
- * @param content
- */
-export const downloadFile = (filename: string, content?: string) => {
-  const element = document.createElement('a')
-
-  if (content) {
-    const type = mime.getType(filename.split('.').pop())
-    element.href = `data:${type};charset=utf-8,` + encodeURIComponent(content)
-  } else {
-    element.href = filename
-  }
-
-  element.download = filename
-  element.click()
-  element.remove()
-}
-
-/**
- * Create a 53 bit hash from a string
- * https://stackoverflow.com/questions/7616461/generate-a-hash-from-string-in-javascript
- * @param val String to hash
- * @param seed Seed (unsigned integer, 32-bit max)
- * @returns A number
- */
-export const hash = (val: string, seed = 0) => {
-  let h1 = 0xdeadbeef ^ seed,
-    h2 = 0x41c6ce57 ^ seed
-  for (let i = 0, ch: number; i < val.length; i++) {
-    ch = val.charCodeAt(i)
-    h1 = Math.imul(h1 ^ ch, 2654435761)
-    h2 = Math.imul(h2 ^ ch, 1597334677)
-  }
-  h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507) ^ Math.imul(h2 ^ (h2 >>> 13), 3266489909)
-  h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507) ^ Math.imul(h1 ^ (h1 >>> 13), 3266489909)
-  return 4294967296 * (2097151 & h2) + (h1 >>> 0)
-}
-
-/**
- * Remove underscore and extension and apply start/title case to a string
- * @param val String to clean
- */
-export const clean = (val: string) => {
-  // apply start/title case
-  return val
-    .replace(/[_-]/g, ' ') // remove underscore and hyphen
-    .replace(/\.[^/.]+$/, '') // remove extension
-    .replace(/\b\w/g, (str) => str.toUpperCase())
-}
-/**
- * Rescale input value based on min/max of old/new scale
- * @param input
- * @param oldMin
- * @param oldMax
- * @param newMin
- * @param newMax
- */
-export const remapInputValue = (input = 0, oldMin = 0, oldMax = 0, newMin = 0, newMax = 255) => {
-  const percent = (input - oldMin) / (oldMax - oldMin)
-  return percent * (newMax - newMin) + newMin
-}
-
-export const getSampleFromInterval = (intervalStart: number, intervalEnd: number, numberOfItems: number) => {
-  const randomSamplesFromInterval = []
-
-  while (randomSamplesFromInterval.length < numberOfItems) {
-    const randomValue = remapInputValue(Math.random(), 0, 1, intervalStart, intervalEnd)
-    randomSamplesFromInterval.push(randomValue)
-  }
-  return randomSamplesFromInterval
-}
-
-export const getIntervalList = (
-  classificationMethod: ClassificationMethodTypes,
-  layerMin: number,
-  layerMax: number,
-  randomSample: number[],
-  numberOfClasses: number,
-) => {
-  let intervalList: number[]
-  if (classificationMethod === ClassificationMethodTypes.NATURAL_BREAK) {
-    intervalList = new Jenks([layerMin, ...randomSample, layerMax], numberOfClasses).naturalBreak().map((element) => {
-      return Number(element.toFixed(2))
-    })
-  } else if ((classificationMethod === ClassificationMethodTypes.LOGARITHMIC && layerMin < 1) || layerMax < 1) {
-    const range = layerMax - layerMin
-    const positive = [layerMin, ...randomSample, layerMax].map((v) => {
-      return remapInputValue(v, layerMin, layerMax, 1, 1 + range)
-    })
-    intervalList = chroma
-      .limits(positive, classificationMethod, numberOfClasses)
-      .map((v) => {
-        return remapInputValue(v, 1, 1 + range, layerMin, layerMax)
-      })
-      .map((element) => {
-        return Number(element.toFixed(2))
-      })
-  } else {
-    intervalList = chroma
-      .limits([layerMin, ...randomSample, layerMax], classificationMethod, numberOfClasses)
-      .map((element) => {
-        return Number(element.toFixed(2))
-      })
-  }
-  return intervalList
-}
-
-/**
- * get tilestats data for specified layer from vector tile
- * @param pbfPath vector tile path. path of 0/0/0.pbf should be used
- * @param layerName layer name on vector tile
- * @returns tilestats information
- */
-export const getVectorInfo = async (pbfPath: string, layerName: string) => {
-  return await fetchUrl(`vectorinfo.json?path=${pbfPath}&layer_name=${layerName}`)
-}
-
-export const groupByN = (n: number, data: any[]) => {
-  const result = []
-  for (let i = 0; i < data.length; i += n) result.push(data.slice(i, i + n))
-  return result
-}
-
-/**
- * Get URL with base 64 encoded signature
- * @param url url with signature
- * @returns url after base 64 encoded signature
- */
-export const getBase64EncodedUrl = (url: string) => {
-  const [base, sign] = url.split('?')
-  return `${base}?${btoa(sign)}`
-}
-
-/**
- * Get active band index number from RasterTileMetadata object
- * if `active_band_no` property is not defined, 0 is returned as default.
- * @param metadata RasterTileMetadata object
- * @returns active band index number
- */
-export const getActiveBandIndex = (metadata: RasterTileMetadata) => {
-  let bandIndex = -1
-  if (metadata.active_band_no) {
-    bandIndex = Object.keys(metadata.band_metadata).indexOf(metadata.active_band_no)
-    for (let i = 0; i < metadata.band_metadata.length; i++) {
-      const band = metadata.band_metadata[i][0]
-      if (band == metadata.active_band_no) {
-        bandIndex = i
-        break
-      }
-    }
-  } else {
-    bandIndex = 0
-    metadata.active_band_no = metadata.band_metadata[bandIndex][0]
-  }
-  return bandIndex
-}
-
-export const generateColorMap = (
-  layer: Layer,
-  layerMin: number,
-  layerMax: number,
-  numberOfClasses: number,
-  classificationMethod: ClassificationMethodTypes,
-  isClassificationMethodEdited: boolean,
-  percentile98: number,
-) => {
-  const colorMap = []
-  if (classificationMethod === ClassificationMethodTypes.LOGARITHMIC) {
-    const randomSample = getSampleFromInterval(layerMin, percentile98, NO_RANDOM_SAMPLING_POINTS)
-    const intervalList = getIntervalList(classificationMethod, layerMin, percentile98, randomSample, numberOfClasses)
-    const scaleColorList = chroma.scale(layer.colorMapName).classes(intervalList)
-    for (let i = 0; i <= numberOfClasses - 2; i++) {
-      const row: IntervalLegendColorMapRow = {
-        index: i,
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore:next-line
-        color: [...scaleColorList(intervalList[i]).rgb(), 255],
-        start:
-          isClassificationMethodEdited == false &&
-          layer.intervals.colorMapRows.length > 0 &&
-          layer.intervals.numberOfClasses === numberOfClasses &&
-          layer.intervals.colorMapRows[i]?.start
-            ? layer.intervals.colorMapRows[i].start
-            : intervalList[i],
-        end:
-          isClassificationMethodEdited == false &&
-          layer.intervals.colorMapRows.length > 0 &&
-          layer.intervals.numberOfClasses === numberOfClasses &&
-          layer.intervals.colorMapRows[i]?.end
-            ? layer.intervals.colorMapRows[i].end
-            : intervalList[i + 1],
-      }
-      colorMap.push(row)
-    }
-    const lastRow: IntervalLegendColorMapRow = {
-      index: numberOfClasses - 1,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore:next-line
-      color: [...scaleColorList(intervalList[numberOfClasses - 1]).rgb(), 255],
-      start: Math.floor(percentile98),
-      end: Math.ceil(layerMax),
-    }
-    colorMap.push(lastRow)
-    const replaceIndex = colorMap[colorMap.length - 2]
-
-    replaceIndex['end'] = Math.floor(percentile98)
-    colorMap.splice(colorMap.length - 2, replaceIndex)
-  } else {
-    const randomSample = getSampleFromInterval(layerMin, layerMax, NO_RANDOM_SAMPLING_POINTS)
-    const intervalList = getIntervalList(classificationMethod, layerMin, layerMax, randomSample, numberOfClasses)
-    const scaleColorList = chroma.scale(layer.colorMapName).classes(intervalList)
-    for (let i = 0; i <= numberOfClasses - 1; i++) {
-      const row: IntervalLegendColorMapRow = {
-        index: i,
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore:next-line
-        color: [...scaleColorList(intervalList[i]).rgb(), 255],
-        start:
-          isClassificationMethodEdited == false &&
-          layer.intervals.colorMapRows.length > 0 &&
-          layer.intervals.numberOfClasses === numberOfClasses &&
-          layer.intervals.colorMapRows[i]?.start
-            ? layer.intervals.colorMapRows[i].start
-            : intervalList[i],
-        end:
-          isClassificationMethodEdited == false &&
-          layer.intervals.colorMapRows.length > 0 &&
-          layer.intervals.numberOfClasses === numberOfClasses &&
-          layer.intervals.colorMapRows[i]?.end
-            ? layer.intervals.colorMapRows[i].end
-            : intervalList[i + 1],
-      }
-      colorMap.push(row)
-    }
-  }
-  return colorMap
-}
