+++ src/lib/helper/getSampleFromInterval.ts
@@ -1,0 +1,11 @@
+import { remapInputValue } from './remapInputValue'
+
+export const getSampleFromInterval = (intervalStart: number, intervalEnd: number, numberOfItems: number) => {
+  const randomSamplesFromInterval = []
+
+  while (randomSamplesFromInterval.length < numberOfItems) {
+    const randomValue = remapInputValue(Math.random(), 0, 1, intervalStart, intervalEnd)
+    randomSamplesFromInterval.push(randomValue)
+  }
+  return randomSamplesFromInterval
+}
