+++ src/lib/helper/hash.ts
@@ -1,0 +1,19 @@
+/**
+ * Create a 53 bit hash from a string
+ * https://stackoverflow.com/questions/7616461/generate-a-hash-from-string-in-javascript
+ * @param val String to hash
+ * @param seed Seed (unsigned integer, 32-bit max)
+ * @returns A number
+ */
+export const hash = (val: string, seed = 0) => {
+  let h1 = 0xdeadbeef ^ seed,
+    h2 = 0x41c6ce57 ^ seed
+  for (let i = 0, ch: number; i < val.length; i++) {
+    ch = val.charCodeAt(i)
+    h1 = Math.imul(h1 ^ ch, 2654435761)
+    h2 = Math.imul(h2 ^ ch, 1597334677)
+  }
+  h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507) ^ Math.imul(h2 ^ (h2 >>> 13), 3266489909)
+  h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507) ^ Math.imul(h1 ^ (h1 >>> 13), 3266489909)
+  return 4294967296 * (2097151 & h2) + (h1 >>> 0)
+}
